V8 bindings gen: Error if FeaturePolicy and SecureContext together.
This check was previously added for OriginTrialEnabled and SecureContext
(r461653). This applies the same check to FeaturePolicy which has the
same limitation (i.e., this *should* work but currently generates broken
code where FeaturePolicy is totally ignored, and therefore must be a
compile-time error for now.

BUG=695123

Review-Url: https://codereview.chromium.org/2802873003
Cr-Original-Commit-Position: refs/heads/master@{#462730}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4dbfdb063d396295597734d850b8f41c20aa50f1