Refactored out the need to use CSSPropertyID to parse shadow properties.
This is pre work to allow the parseSingleValue method of the
CSSPropertyTextShadow and CSSPropertyBoxShadow properties to be
implemented in separate APIs.

This patch also makes the arguments passed to consumeShadow consistent
with those passed to parseSingleShadow and removes property specific
naming in those arguments.

BUG=668012

Review-Url: https://codereview.chromium.org/2796803002
Cr-Original-Commit-Position: refs/heads/master@{#462808}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7586cc3f2ec274b36870e7b70e3acc034f0dbd2f