bindings: Skip undefined property descriptors when creating records.
We were being too aggressive with our DCHECK(desc->IsObject()). The spec
says (and we even mention it in a comment) that a call to
Object.getOwnPropertyDescriptor can return undefined in addition to a proper
object descriptor.

We were only considering the latter case, and crashing on the DCHECK if we
ever received an undefined descriptor, which can happen when a Proxy is used
in JS.

This is also covered by external/wpt/fetch/api/headers/headers-record.html.

BUG=707365
R=bashi@chromium.org,haraken@chromium.org,yukishiino@chromium.org

Review-Url: https://codereview.chromium.org/2801003005
Cr-Original-Commit-Position: refs/heads/master@{#462850}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9e1372a28311ec4b7fee0ae76eba7301069e7c6a