Remove geolocation-api/success-clear-watch.html test from ASAN exceptions file.
In Jan 20015, this test case on a ASAN run caused timeouts. An exception was
added and life moved on. Re-running this tests locally does not appear to cause
any timeout, and so we're going to re-enable the exception to see if there still
exists a problem.

BUG=446661

Review-Url: https://codereview.chromium.org/2799743007
Cr-Original-Commit-Position: refs/heads/master@{#462945}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b726dba11c90d226f70c9dae4de30c002c066131