Re-enable selectors4 tests from WPT repository
As agreed on https://chromium-review.googlesource.com/465440
the tests that are considered manual due to the "interact" flag,
but that the feature hasn't been implemented yet ":focus-within"
are skipped for now.

There's one manual test "hover-001-manual.html"
that is also skipped in this patch.

BUG=706118

Review-Url: https://codereview.chromium.org/2803213002
Cr-Original-Commit-Position: refs/heads/master@{#462982}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 333a624229a859b3388e4c6af0f29197709991a2