Avoid duplicate functions/code in core/editing: endTag
While experimenting with unity builds I encountered a few duplicate
symbols and functions in core/editing. This patch renames, moves
and unifies them.

elementCannotHaveEndTag is a utility function used in serializers
and since it is used in multiple places, and MarkupFormatter is
not a good place for it, let us put it in EditingUtilities.

BUG=708949
R=yosin@chromium.org

Review-Url: https://codereview.chromium.org/2804943002
Cr-Original-Commit-Position: refs/heads/master@{#462994}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a7159abf14936edb1b8782dc06253c1ff91b9000