Rename caretDisplayItemClient() to getCaretDisplayItemClient().
The rename is needed to avoid a naming collision after changing from
Blink to Chromium naming style.  Right now we have a
|CaretDisplayItemClient| type and a |caretDisplayItemClient| accessor
method in the CaretDisplayItemClientTest class (differing by case of the
first character); after a naive rename by the rewrite_to_chrome_style
tool we would end up with |CaretDisplayItemClient| being the name of
both the type and the accessor method (with both living in the same
namespace).

Prepending a "get" prefix to the name of the accessor method is the
workaround that fits into the guidance on the recommended
post-Blink-to-Chromium-rename style suggested by esprehn@ in
https://crbug.com/582312#c17:
- Getters favor not using "Get", ex. FirstChild()
- Unless the type name conflicts, in which case you can either rename
  the type if it's easy and makes sense, or add "Get", ex. GetContext().

BUG=582312
TBR=yosin@chromium.org

Review-Url: https://codereview.chromium.org/2803363002
Cr-Original-Commit-Position: refs/heads/master@{#463051}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 00daa6ffb9302d37c20324ab4f36485969b63a5a