Verify that constant vectors aren't invalidated during iteration.
Clearing a heap vector while it is being iterated and on the stack is
unsafe, as it promptly releases the backing store.

Add CHECK()s to verify that this doesn't happen for a crasher involving
MediaQuerySets that's proving hard to pindown, but there are some
suggestions that the heap vector contents of MediaQuerySet::m_queries
is being mutated.

R=
BUG=699269

Review-Url: https://codereview.chromium.org/2806003002
Cr-Original-Commit-Position: refs/heads/master@{#463124}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b2a0ce3e8a9b3e561c67cf0a2772c4567567b9f6