SELECT popup: Make OPTIONs taller for touchscreen.
On Windows and Linux:
If a touchscreen is connected, OPTIONs in SELECT popups have 8px paddings and at
least 24px height in order to make it touch-friendly.

On CrOS:
Apply the taller OPTIONs regardless of existence of touchscreen.

On macOS and Android:
This CL has no effect.

BUG=703971

Review-Url: https://codereview.chromium.org/2810583002
Cr-Original-Commit-Position: refs/heads/master@{#463194}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 112094796a99ff35f551467912814d08935b6055