Cache API tests: prepopulate cache in deterministic order
Chrome and Firefox differ in the order in which cache keys() are
returned. Chrome orders by according to when the put()s were issued.
Firefox orders by when the body is complete. The test helper
prepopulated_cache_test did not guarantee that these matched, leading
to the tests being flaky in Firefox. This change tweaks the helper so
that the put()s are processed serially so that the order is
deterministic for both.

Spec issue: https://github.com/w3c/ServiceWorker/issues/823
BUG=655479

Review-Url: https://codereview.chromium.org/2806793002
Cr-Original-Commit-Position: refs/heads/master@{#463195}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5136241948671b538965c783e8aee27e88037aa9