bindings: Call member.includes_for_type() for container types too.
Follow-up to commit 8552274ebe ("bindings: Correctly expand all class/header
dependencies in unions"), which caused member.includes_for_type() not to be
called when a nested union member is processed.

We now call it in all cases, even for container types such as records,
sequences and unions, otherwise we risk skipping includes that come from the
type (e.g. IDLRecord) itself. This does not happen at the moment, but may in
the future, such as when we start converting sequences via IDLSequence and
NativeValueTraits::NativeValue().

BUG=690428
R=bashi@chromium.org,haraken@chromium.org,yukishiino@chromium.org

Review-Url: https://codereview.chromium.org/2806343002
Cr-Original-Commit-Position: refs/heads/master@{#463268}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0338b35867900b81de5e6960eec0572f152974b7