Tests of ctrl-click can't postMessage to opener - using custom text instead.
This CL fixes/refactors two layout tests that used to depend on ability
to find the opener via window.open, even in case of ctrl-click or
shift-click.  This dependency is incorrect and will be broken once
https://crbug.com/658386 is fixed.

The tests had to be refactored so that they don't depend on ability to
communicate back with the main test window.  After this CL the tests
report their results via testRunner.setCustomTextOutput (rather than
sending the results back to the main window using window.postMessage).

In the 2 tests here, the custom text dump is being set from a secondary
test window - the CL also needs to relax the condition that limited
propagation of LayoutTestRuntimeFlags to the main test window.  The
intent of the original condition should have covered both secondary
and main windows (i.e. the intent was to just check if the renderer
has received a LayoutTestRuntimeFlags).  It seems desirable overall
to also propagate LayoutTestRuntimeFlags changes originating from
secondary test windows.

BUG=658386

Review-Url: https://codereview.chromium.org/2689483007
Cr-Original-Commit-Position: refs/heads/master@{#463400}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 69e056006f20f2973eafb2e95b7ea4395f3065db