CHECK that SuspendableScriptExecutor is always created with a valid ScriptState.
Something is going awry, and we suspect that it's not having a valid state
when we go to execute it. It seems reasonable to check whether that is happening
as we prepare to run the script, or whether we already had a bad ScriptState
when the object was created.

BUG=693014

Review-Url: https://codereview.chromium.org/2811823002
Cr-Original-Commit-Position: refs/heads/master@{#463463}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7e550b9dbb90d9b72e51778a937ee5b03ef7f98f