Clean up DocumentMarkerController::copyMarkers() (now it's moveMarkers())
This method is kind of weird since it leaves the markers in the source list but
sometimes changes one of them (if it spans a node boundary). In this CL, I'm
cleaning it up so it removes the markers from the source list after adding them
to the destination list.

BUG=707867

Review-Url: https://codereview.chromium.org/2810463002
Cr-Original-Commit-Position: refs/heads/master@{#463509}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0dea0097ea24907ed4c0f30fb4180191f378f00e