Allow ComputedStyleBase property fields to be storage_only.
Currently the 'storage_only' template can only be applied to nonproperty
fields because 'storage_only' is not a valid value for 'template' in
CSSProperties.json5. However, there are cases when we want a property
field to be 'storage_only' (e.g. for color related properties whose
getters are very complex).

This patch adds 'storage_only' as a valid template option in
CSSProperties.json5.

BUG=628043

Review-Url: https://codereview.chromium.org/2801523002
Cr-Original-Commit-Position: refs/heads/master@{#463874}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 83e9ff4b5ed44bf5c2196f3bdfdde63a07c6072e