Set plugin focus by implementing HTMLPlugInElement::SetFocused.
Set plugin focus in HTMLPlugInElement::SetFocused rather than an
explicit check and call in Document::SetFocusedElement.
This causes plugins to have DidFocusChange executed twice
(true/false), if the DOM plugin element changes focus in
onblur or onfocus, but the plugin does remain in a consistent
state as to whether it has focus or not.

BUG=637460

Review-Url: https://codereview.chromium.org/2810593002
Cr-Original-Commit-Position: refs/heads/master@{#463917}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e7aa842423521bcc232e2f0c3070196038f92f25