createPeriodicWave parameters are sequence<float>
Change the parameter types for createPeriodicWave from Float32Array to
sequence<float>.  this makes createPeriodicWave signal errors if
non-finite Fourier coefficients are used and matches what the
PeriodicWave constructor behavior.

Also periodic-lengths.html was actually generating Infinity for some
coefficient values, so the test is updated to correct for that.  And
the selected waveform is changed from square to sawtooth because the
Fourier series for a sawtooth is non-zero everywhere.  This fixes the
problem where the wave with 8192 coefficients was the same as for
8191.  Which has to be true because a square wave only has non-zero
coefficients for the odd terms, so the arrays for 8192 and 8191 were
the same.  With a sawtooth, these are now different.

Add new test to verify that we throw errors if non-finite values are
used for the Fourier coefficients.

BUG=709512
TEST=periodicwave-lengths.html, periodicwave-exceptions.html

Review-Url: https://codereview.chromium.org/2807673002
Cr-Original-Commit-Position: refs/heads/master@{#464441}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 445c9544f743c5fb1dee4ba1594360bb10366f92