[Bindings] Make maplike<> and setlike<> imply a readonly size attribute.
This required implementing one for RTCStatsReport, which is presently missing
it.

This also replaces the size attribute on FontFaceSet, MIDIInputMap and
MIDIOutputMap with one generated according to the WebIDL rules for maplike<>
and setlike<>.

Because the spec specifies that this property is not enumerable, this has the
side effect of making it non-enumerable on the interfaces that already define
it explicitly.

BUG=697121
TBR=tommyw@chromium.org

Review-Url: https://codereview.chromium.org/2813023002
Cr-Original-Commit-Position: refs/heads/master@{#464501}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3ca2419cbb0dc16043535a1ec1e04b4977e725c1