setValueCurveAtTime takes sequence<float> for curve
Instead of a Float32Array, the curve parameter for an
AudioParam.setValueCurveAtTime method is now a sequence<float>.  This
also implies that only finite values for the curve.

This change is (mostly) backward compatible since Float32Array is a
sequence<float>.

Added a test to verify that we throw an error if the curve contains a
non-finite value.

BUG=708306
TEST=AudioParam/audioparam-exceptional-values.html

Review-Url: https://codereview.chromium.org/2789883002
Cr-Original-Commit-Position: refs/heads/master@{#464504}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e78f7c1cc5587c0a09ca68b9519a45b1c3238ece